Refactoring

+++ src/MetroFramework5/Animation/ExpandAnimation.cs
@@ -48,7 +48,7 @@ namespace MetroFramework5.Animation
         private int DoExpandAnimation(int startSize, int targetSize)
         {
             float t = (float)counter - startTime;
-            float b = (float)startSize;
+            float b = startSize;
             float c = (float)targetSize - startSize;
             float d = (float)targetTime - startTime;
