+++ src/MetroFramework5/Animation/MoveAnimation.cs
@@ -48,7 +48,7 @@ namespace MetroFramework5.Animation
         private int DoMoveAnimation(int startPos, int targetPos)
         {
             float t = (float)counter - startTime;
-            float b = (float)startPos;
+            float b = startPos;
             float c = (float)targetPos - startPos;
             float d = (float)targetTime - startTime;
